Crate syscall_logger

Attribute Macros

log_syscall

This macro wraps a syscall handler function by renaming the original function and making a new function with the original name that calls the original function. When the syscall handler function is called, it will log the syscall if syscall logging is enabled in Shadow.